perf(decode): pack HUF decode table as u16 (donor HUF_DEltX1 layout)
Donor `huf_decompress.c` defines the HUF decoding table entry as
`HUF_DEltX1 = {BYTE byte; BYTE nbBits}` — exactly 2 bytes per state
index. The inner loop reads `dtable[index]` as a 16-bit scalar load
and splits the low byte (symbol) from the high byte (nbBits).

Our `packed_decode: Vec<u32>` used a 32-bit slot per state index
even though only the low 16 bits carried payload. At
`max_num_bits = 11` (zstd spec ceiling) that doubled the HUF table
from donor's 4 KiB to 8 KiB. On poorly-compressed inputs (L-7 Fast
in particular) the HUF decode reads 1 entry per output literal byte,
the literal buffer is ~100 KiB per block, and the table competes
with the output buffer for L1d residency. Halving the table
footprint to match donor restores the cache hit rate on those
workloads.

Changes:

- `HuffmanTable::packed_decode: Vec<u32>` → `Vec<u16>`. Build site
  packs `u16::from(symbol) | (u16::from(num_bits) << 8)` — same
  layout as donor `HUF_DEltX1` (symbol in low byte, nbBits in high
  byte).
- `run_4stream_decode_loop`'s `packed: &[u32]` parameter → `&[u16]`.
  The burst body's `(entry & 0xFF) as u8` + `(entry >> 8) & 0xFF`
  pattern is unchanged in shape since the low/high byte split is
  identical between u32 and u16.
- AVX2 kernel: dropped `_mm_i32gather_epi32`. Donor never uses
  gather — `HUF_4X1_DECODE_SYMBOL` reads scalar entries through
  multiple load ports. Match donor: four scalar u16 lookups.
- VBMI2 kernel: still uses `_mm_maskz_compress_epi8` to compact
  the 4 entries down to {[symbols], [bits]} byte tuples, but loads
  via scalar u16 reads into a packed u64 instead of the prior u32
  lane-set.
- NEON kernel: dropped `vld1q_u32`/`vandq_u32`/`vshrq_n_u32` round
  trip. M1's 4-way load pipeline executes four scalar u16 reads in
  fewer cycles than the SIMD pack-unpack on this 4-element batch.
- SVE kernel: dropped the `ld1w`/`whilelt`/`p0.s` inline-asm
  4-lane unpack. Scalar u16 reads + byte split — donor parity.
- `decode4_symbols_and_num_bits_avx2` retains its `#[target_feature]`
  guard for the kernel-dispatch surface but its body is now scalar.
- Test fixture in `huff0_decoder.rs`: `u32::from` → `u16::from` for
  the manually-built `packed_decode` vector.
- Drop now-unused `vandq_u32` / `vdupq_n_u32` / `vld1q_u32` /
  `vshrq_n_u32` / `vst1q_u32` / `asm` imports.
- Drive-by: tighten the `block_splitter_decision_for_bench` cfg
  gate from `any(test, feature = "bench_internals")` to just
  `feature = "bench_internals"`. The previous form fired
  dead-code warnings under `cargo clippy --all-targets` without
  the feature.

Tests: 577/577 nextest pass, clippy clean. The L-7 Fast decompress
hot path (HUF 4-stream burst body + SIMD fallback kernels) is the
target — bench validation requires CI-equivalent x86_64 runners
since the regression presents most acutely on cache-pressured
shared-machine workloads.

Part of #178.

+++ zstd/src/decoding/literals_section_decoder.rs
@@ -462,7 +462,7 @@ unsafe fn run_4stream_decode_loop<K: HufKernel>(
     decoders: &mut [HuffmanDecoder<'_>; 4],
     brs: &mut [BitReaderReversed<'_>; 4],
     target: &mut [u8],
-    packed: &[u32],
+    packed: &[u16],
     cursors: &mut [usize; 4],
     bounds: &LoopBounds,
 ) {
